requirements.txt
Flask==0.12.2
Flask-MQTT==0.0.11
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
click==6.7
imutils==0.4.5
itsdangerous==0.24
numpy==1.14.0
paho-mqtt==1.3.1
typing==3.6.4
